+++ lemon-biz/src/main/java/com/lemon/lemonbiz/mail/controller/MailController.java
@@ -178,19 +178,6 @@ public class MailController {
 		 // *mail 앞에 / 빼먹지 말 것*
 		 return "redirect:/mail/sendMail";
 		} 
-	
-	@RequestMapping("/getCountNoReadMail")
-	public ResponseEntity<?> getTodayCount(@RequestParam HashMap<Object,Object> params) {
-		
-		System.out.println("@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@");
-		System.out.println("params = " + params);
-		
-		int num = mailService.getCountNoReadMail(params);
-		
-		System.out.println("num = " + num);
-
-		return new ResponseEntity<>(num,HttpStatus.OK);		
-	}
 
 }
