+++ lemon-biz/src/main/java/com/lemon/lemonbiz/mail/model/dao/MailDAO.java
@@ -34,5 +34,5 @@ public interface MailDAO {
 
 	Member selectMyInfo(Member loginMember);
 
-	int getCountNoReadMail(HashMap<Object, Object> params);
+	
 }
