+++ lemon-biz/src/main/resources/mapper/mail/mail-mapper.xml
@@ -179,16 +179,7 @@
 		WHERE KEY = #{ key }
 	</select>
 	
-	<select id="getCountNoReadMail" resultType="int">
-		select 
-			count(*)
-		from
-			om
-		where 
-			mem_id = #{ memberId }
-		and 
-			read_count = '0'
-	</select>
+	
 </mapper>
